Update Partner-overview.md
replicates on-premises Exchange mailbox data into Office 365 for access by Workplace Analytics.  Implemented as a managed service, it can be deployed on-premises or in Azure and features encrypted end-to-end transfers, auditing and chain-of-custody for optimal security and compliance.  WPA-SYNC also has a range of options for controlling operations and ensuring scalability in large/complex environments.  It uses SaaS with Exchange Web Services and requires access to the customer's domain. To learn more about this solution,